rpc-parser: Move wrapper in to parser.
